-1- In 2 migrations changed integer to tyny integer

+++ database/migrations/2018_05_03_183850_create_cards_table.php
@@ -15,7 +15,7 @@ class CreateCardsTable extends Migration
     {
         Schema::create('cards', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedInteger('type_id');
+            $table->unsignedSmallInteger('type_id');
             $table->string('category', 20);
             $table->string('image', 100);
         });
